tools/gmail/pyproject.toml
[project]
name = "gmail-collector"
version = "0.1.0"
description = "Gmail content collector tool"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.100.0",
    "google-auth-httplib2>=0.1.1",
    "google-auth-oauthlib>=1.1.0",
    "click>=8.0.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]

[project.scripts]
gmail-collect = "tools.gmail.src.collectors.gmail.cli:main"
gmail-send = "tools.gmail.src.collectors.gmail.send_cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
